10132024 Commit
Continued modifications to allow the program to be used to connect to both the camera and the robot controller.

Currently Network Scanning for active controllers is working although it keeps adding the found controllers even when it already exists. Need to add an if exists.

Connect to controller currently hangs the program