Edited multi frame capture code after conversion to SendCommand.
Removed all WSL deffinitions. Future use of WSL will be made via ssh to a windows shell. This will preclude the need to develop two methods for each function. openSSH can be installed on the WSL instance and connect to that to run the gphoto2 commands on the local linux shell.